+++ R/normaliseData.R
@@ -1,7 +1,7 @@
 
 #' Apply normalisation on a data frame using specific Lambda value
 #'
-#' @param data a data frame containing the variables.
+#' @param dat a data frame containing the variables.
 #' @param lambdas a numeric vector from the \link[MASS]{boxcox} function
 #'
 #' @return a data frame
@@ -11,8 +11,8 @@
 #' \dontrun{normaliseData(mtcars, getLambda(mtcars, parallel = FALSE))}
 #' normaliseData(mtcars, getLambda(mtcars, parallel = FALSE))
 normaliseData <-
-  function(data, lambdas){
-    stopifnot(is.data.frame(data) | is.numeric(lambdas))
+  function(dat, lambdas){
+    stopifnot(is.data.frame(dat) | is.numeric(lambdas))
 
-    as.data.frame(apply(rbind(lambdas, data), 2, function(x) normalise(x[-1], x[1])))
+    as.data.frame(apply(rbind(lambdas, dat), 2, function(x) normalise(x[-1], x[1])))
   }